+++ src/preset/bidirectional.rs
@@ -33,57 +33,12 @@ impl<P: Bundle, N: Bundle> SpawnableList<BindingOf> for Bidirectional<P, N> {
 }
 
 impl Bidirectional<Binding, Binding> {
-    /// Maps A and D keys as 1-dimensional input.
+    /// Maps 2 bindings as 1-dimensional input.
     #[must_use]
-    pub fn ad_keys() -> Self {
+    pub fn new(positive: impl Into<Binding>, negative: impl Into<Binding>) -> Self {
         Self {
-            positive: KeyCode::KeyD.into(),
-            negative: KeyCode::KeyA.into(),
-        }
-    }
-
-    /// Maps W and S keys as 1-dimensional input.
-    #[must_use]
-    pub fn ws_keys() -> Self {
-        Self {
-            positive: KeyCode::KeyW.into(),
-            negative: KeyCode::KeyS.into(),
-        }
-    }
-
-    /// Maps left and right keyboard arrow keys as 1-dimensional input.
-    #[must_use]
-    pub fn left_right_arrow() -> Self {
-        Self {
-            positive: KeyCode::ArrowRight.into(),
-            negative: KeyCode::ArrowLeft.into(),
-        }
-    }
-
-    /// Maps up and down keyboard arrow keys as 1-dimensional input.
-    #[must_use]
-    pub fn up_down_arrow() -> Self {
-        Self {
-            positive: KeyCode::ArrowUp.into(),
-            negative: KeyCode::ArrowDown.into(),
-        }
-    }
-
-    /// Maps up and down D-pad buttons as 1-dimensional horizontal input.
-    #[must_use]
-    pub fn left_right_dpad() -> Self {
-        Self {
-            positive: GamepadButton::DPadRight.into(),
-            negative: GamepadButton::DPadLeft.into(),
-        }
-    }
-
-    /// Maps left and right D-pad buttons as 1-dimensional horizontal input.
-    #[must_use]
-    pub fn up_down_dpad() -> Self {
-        Self {
-            positive: GamepadButton::DPadUp.into(),
-            negative: GamepadButton::DPadDown.into(),
+            positive: positive.into(),
+            negative: negative.into(),
         }
     }